refactor: 코멘트 작성 시 로직 service 및 메서드 분리
